Revert "warningの削除(2回目)"
This reverts commit 6200363fe77d88e837a67f33a8afaec82d5022ce.

+++ Game/UI/Header/Gauge.h
@@ -32,7 +32,7 @@ public:
 	// ゲージ量を設定
 	void SetMaxValue(float max)		{ m_value = max; }
 	// ゲージ量の取得
-	float GetValue()				{ return m_value; }
+	int GetValue()					{ return m_value; }
 
 private:
 	// スプライトバッチ
